+++ CHANGELOG.md
@@ -58,6 +58,7 @@ You can find its changes [documented below](#070---2021-01-01).
 - Add #[data(eq)] shorthand attribute for Data derive macro ([#1884] by [@Maan2003])
 - X11: detect keyboard layout ([#1779] by [@Maan2003])
 - WindowDesc::with_config ([#1929] by [@Maan2003])
+- `Notification::route` ([#1978] by [@xarvic])
 
 ### Changed
 
@@ -793,8 +794,9 @@ Last release without a changelog :(
 [#1919]: https://github.com/linebender/druid/pull/1919
 [#1929]: https://github.com/linebender/druid/pull/1929
 [#1947]: https://github.com/linebender/druid/pull/1947
-[#1967]: https://github.com/linebender/druid/pull/1967
 [#1953]: https://github.com/linebender/druid/pull/1953
+[#1967]: https://github.com/linebender/druid/pull/1967
+[#1978]: https://github.com/linebender/druid/pull/1978
 
 [Unreleased]: https://github.com/linebender/druid/compare/v0.7.0...master
 [0.7.0]: https://github.com/linebender/druid/compare/v0.6.0...v0.7.0